move trip details functions from data.js to component file
