+++ src/js/components/trip-details.js
@@ -1,16 +1,31 @@
 import {createElement} from '../utils.js';
 
-/*export const renderTripInfo = ({title, dates}) =>
-  `<div class="trip-info__main">
-    <h1 class="trip-info__title">${title()}</h1>
-    <p class="trip-info__dates">${dates()}</p>
-  </div>`.trim();*/
+const getTripTitle = (eventsList) => {
+  let tripRoute = [];
+  eventsList.forEach((event) => {
+    tripRoute.push(event.destination);
+  });
+  if (tripRoute.length > 3) {
+    let tripTitle = [];
+    tripTitle.push(tripRoute.shift(), tripRoute.pop());
+    return tripTitle.join(` — ... — `);
+  } else {
+    return tripRoute.join(` — `);
+  }
+};
+
+const getTripDates = (eventsList) => {
+  let tripDates = [...Array(2)];
+  tripDates[0] = new Date(eventsList[0].dateTime.dateStart).toString().slice(4, 10);
+  tripDates[1] = new Date(eventsList[eventsList.length - 1].dateTime.dateEnd()).toString().slice(4, 10);
+  return tripDates.join(`&nbsp;—&nbsp;`);
+};
 
 export class TripDetails {
-  constructor ({title, dates}) {
+  constructor(eventsList) {
     this._element = null;
-    this._title = title();
-    this._dates = dates();
+    this._title = getTripTitle(eventsList);
+    this._dates = getTripDates(eventsList);
   }
 
   getElement() {
